move cmd files to separate folders
