+++ pkg/cmd-utils/init.go
@@ -1,0 +1,3 @@
+package cmdutils
+
+var RecipeDir string
